Fixed my mistakes + comments
So i was looking at the functions and really the only difference between
my find and sam's increment frequency is that sam's will add 1 to the
frequency.
If my find function finds the word in the list of used words; sam should
yours just simply increment frequency by 1?  It's pointless to search
for the word twice unless I'm not getting this correctly.. ?