Added a shit ton of doxygen *Needs work*
Added doxygen to all of my functions I wrote and also added the main
header. The main header still needs some work as I just did the simple
stuff and you guys need to add your names to it and we need to write a
summary of the program. Also need to look into how to doxygen a class
because currently I only added doxygen to the structure.